move query mapper tocore
